+++ components/core/ui/navigation/src/main/java/com/flipperdevices/core/ui/navigation/FeatureScreenRootRoute.kt
@@ -1,6 +1,0 @@
-package com.flipperdevices.core.ui.navigation
-
-enum class FeatureScreenRootRoute {
-    WEAR_KEYS_LIST,
-    WEAR_EMULATE,
-}