+++ components/core/ui/theme/src/main/kotlin/com/flipperdevices/core/ui/theme/composable/FlipperPalletProvider.kt
@@ -7,7 +7,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.graphics.Color
 import com.flipperdevices.core.preference.pb.SelectedTheme
 import com.flipperdevices.core.ui.theme.viewmodel.ThemeViewModel
-import tangle.viewmodel.compose.tangleViewModel
 
 /**
  * Please, use LocalPallet instead
@@ -33,9 +32,9 @@ fun getThemedFlipperPallet(
 }
 
 @Composable
-fun isLight(
-    systemIsDark: Boolean = isSystemInDarkTheme(),
-    themeViewModel: ThemeViewModel = tangleViewModel()
+internal fun isLight(
+    themeViewModel: ThemeViewModel,
+    systemIsDark: Boolean = isSystemInDarkTheme()
 ): Boolean {
     val theme by themeViewModel.getAppTheme().collectAsState()