+++ components/faphub/main/impl/src/main/java/com/flipperdevices/main/impl/composable/ComposableFapHubMainScreen.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -14,8 +13,6 @@ import com.flipperdevices.core.ui.ktx.SetUpStatusBarColor
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.main.impl.composable.switch.ComposableFapHubNewSwitch
 import com.flipperdevices.main.impl.model.FapHubTabEnum
-import com.flipperdevices.main.impl.viewmodel.MainViewModel
-import tangle.viewmodel.compose.tangleViewModel
 
 @Composable
 fun ComposableFapHubMainScreen(
@@ -24,11 +21,10 @@ fun ComposableFapHubMainScreen(
     installedTabComposable: @Composable () -> Unit,
     onOpenSearch: () -> Unit,
     installedNotificationCount: Int,
+    onSelect: (FapHubTabEnum) -> Unit,
+    selectedTab: FapHubTabEnum,
     modifier: Modifier = Modifier,
-    mainViewModel: MainViewModel = tangleViewModel()
 ) {
-    val selectedTab by mainViewModel.getTabFlow().collectAsState()
-
     Column(
         modifier = modifier,
         horizontalAlignment = Alignment.CenterHorizontally
@@ -36,7 +32,7 @@ fun ComposableFapHubMainScreen(
         SetUpStatusBarColor(LocalPallet.current.accent, darkIcon = true)
         ComposableFapHubNewSwitch(
             fapHubTabEnum = selectedTab,
-            onSelect = mainViewModel::onSelectTab,
+            onSelect = onSelect,
             installedNotificationCount = installedNotificationCount,
             onBack = onBack,
             onEndClick = onOpenSearch