+++ components/faphub/uninstallbutton/impl/src/main/kotlin/com/flipperdevices/faphub/uninstallbutton/impl/api/FapUninstallImpl.kt
@@ -6,13 +6,18 @@ import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.faphub.appcard.composable.ComposableAppDialogBox
 import com.flipperdevices.faphub.dao.api.model.FapItem
 import com.flipperdevices.faphub.dao.api.model.FapItemShort
+import com.flipperdevices.faphub.installation.queue.api.FapInstallationQueueApi
+import com.flipperdevices.faphub.installation.queue.api.model.FapActionRequest
 import com.flipperdevices.faphub.uninstallbutton.api.FapUninstallApi
 import com.flipperdevices.faphub.uninstallbutton.impl.composable.ComposableFapUninstall
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
+import javax.inject.Provider
 
 @ContributesBinding(AppGraph::class, FapUninstallApi::class)
-class FapUninstallImpl @Inject constructor() : FapUninstallApi {
+class FapUninstallImpl @Inject constructor(
+    private val queueApiProvider: Provider<FapInstallationQueueApi>
+) : FapUninstallApi {
     @Composable
     override fun ComposableFapUninstallButton(
         modifier: Modifier,
@@ -21,8 +26,10 @@ class FapUninstallImpl @Inject constructor() : FapUninstallApi {
     ) {
         ComposableFapUninstall(
             modifier = modifier,
-            applicationUid = applicationUid,
-            dialogAppBox = dialogAppBox
+            dialogAppBox = dialogAppBox,
+            onDelete = {
+                queueApiProvider.get().enqueue(FapActionRequest.Delete(applicationUid))
+            }
         )
     }