+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/api/KeyScreenViewDecomposeComponentImpl.kt
@@ -1,9 +1,13 @@
 package com.flipperdevices.keyscreen.impl.api
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.decompose.router.stack.StackNavigation
 import com.arkivanov.decompose.router.stack.push
+import com.arkivanov.essenty.backhandler.BackCallback
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
 import com.flipperdevices.bridge.synchronization.api.SynchronizationUiApi
 import com.flipperdevices.core.ui.ktx.viewModelWithFactory
@@ -18,6 +22,8 @@ import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.update
 
 @Suppress("LongParameterList")
 class KeyScreenViewDecomposeComponentImpl @AssistedInject constructor(
@@ -31,13 +37,33 @@ class KeyScreenViewDecomposeComponentImpl @AssistedInject constructor(
     private val nfcEditor: NfcEditorApi,
     private val keyEmulateApi: KeyEmulateApi,
 ) : DecomposeComponent, ComponentContext by componentContext {
+    private val isBackPressHandledFlow = MutableStateFlow(false)
+    private val backCallback = BackCallback(false) { isBackPressHandledFlow.update { true } }
+
+    init {
+        backHandler.register(backCallback)
+    }
+
     @Composable
     @Suppress("NonSkippableComposable")
     override fun Render() {
         val viewModel: KeyScreenViewModel = viewModelWithFactory(key = keyPath.toString()) {
             keyScreenViewModelFactory(keyPath)
         }
-        shareBottomApi.ComposableShareBottomSheet(viewModel.keyPath) { onShare ->
+        shareBottomApi.ComposableShareBottomSheet(
+            flipperKeyPath = viewModel.keyPath,
+            onSheetStateVisible = { isShown, onClose ->
+                val isBackPressHandled by isBackPressHandledFlow.collectAsState()
+                backCallback.isEnabled = isShown
+
+                LaunchedEffect(isBackPressHandled) {
+                    if (isBackPressHandled) {
+                        onClose()
+                        isBackPressHandledFlow.emit(false)
+                    }
+                }
+            }
+        ) { onShare ->
             ComposableKeyScreen(
                 viewModel = viewModel,
                 synchronizationUiApi = synchronizationUiApi,