+++ components/wearable/emulate/wear/api/src/main/java/com/flipperdevices/wearable/emulate/api/WearEmulateDecomposeComponent.kt
@@ -1,0 +1,16 @@
+package com.flipperdevices.wearable.emulate.api
+
+import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
+import com.flipperdevices.ui.decompose.DecomposeComponent
+import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
+
+interface WearEmulateDecomposeComponent : DecomposeComponent {
+    fun interface Factory {
+        operator fun invoke(
+            componentContext: ComponentContext,
+            flipperKeyPath: FlipperKeyPath,
+            onBack: DecomposeOnBackParameter
+        ): WearEmulateDecomposeComponent
+    }
+}