+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 - [Refactor] Migrate archive screen to decompose
 - [Refactor] Migrate bottombar to decompose
 - [Refactor] Migrate root screens to decompose
-- [Refactor] Add deeplink handle
+- [Refactor] Add deeplink handle to decompose
+- [Refactor] Migrate wearos to decompose
+- [Refactor] Remove tangle and jetpack navigation deps
 - [Feature] Button vibration in ScreenStreaming
 
 # 1.6.7